BOAC-5860, note-templates tab on Peer Advising Mgmt Dashboard
